Refactor: move module exports generation from adapter to compiler layer
Module exports (exported constants/functions) are pure JS output, not
adapter-specific template concerns. This change extracts the shared
logic into packages/jsx/src/module-exports.ts, adds structured
TemplateSections to AdapterOutput, and assembles the final output in
compiler.ts — eliminating duplication across HonoAdapter and TestAdapter
and removing fragile string-parsing in the multi-component combiner.
