+++ packages/jsx/src/adapters/test-adapter.ts
@@ -17,7 +17,7 @@ import type {
   IRFragment,
   ParamInfo,
 } from '../types'
-import { type AdapterOutput, BaseAdapter } from './interface'
+import { type AdapterOutput, type TemplateSections, BaseAdapter } from './interface'
 
 export class TestAdapter extends BaseAdapter {
   name = 'test'
@@ -36,11 +36,19 @@ export class TestAdapter extends BaseAdapter {
       ? `\nexport default ${this.componentName}`
       : ''
 
-    const moduleExports = this.generateModuleExports(ir)
-    const template = [imports, types, moduleExports, component].filter(Boolean).join('\n\n') + defaultExport
+    const sections: TemplateSections = {
+      imports,
+      types: types || '',
+      component,
+      defaultExport,
+    }
+
+    // Assemble template for backward compat
+    const template = [imports, types, component].filter(Boolean).join('\n\n') + defaultExport
 
     return {
       template,
+      sections,
       types: types || undefined,
       extension: this.extension,
     }
@@ -102,39 +110,6 @@ export class TestAdapter extends BaseAdapter {
       lines.push('  __instanceId?: string')
       lines.push('  __bfScope?: string')
       lines.push('}')
-    }
-
-    return lines.length > 0 ? lines.join('\n') : null
-  }
-
-  private generateModuleExports(ir: ComponentIR): string | null {
-    const lines: string[] = []
-
-    for (const constant of ir.metadata.localConstants) {
-      if (!constant.isExported) continue
-      const keyword = constant.declarationKind ?? 'const'
-      if (!constant.value) {
-        lines.push(`export ${keyword} ${constant.name}`)
-        continue
-      }
-      const value = constant.value.trim()
-      const isArrowFunc =
-        value.startsWith('async (') ||
-        value.startsWith('function') ||
-        /^\w+\s*=>/.test(value) ||
-        /^\([^)]*\)\s*=>/.test(value)
-
-      if (isArrowFunc) {
-        lines.push(`export ${keyword} ${constant.name} = () => {}`)
-      } else {
-        lines.push(`export ${keyword} ${constant.name} = ${constant.value}`)
-      }
-    }
-
-    for (const func of ir.metadata.localFunctions) {
-      if (!func.isExported) continue
-      const params = func.params.map((p) => p.name).join(', ')
-      lines.push(`export function ${func.name}(${params}) ${func.body}`)
     }
 
     return lines.length > 0 ? lines.join('\n') : null
